LayoutTests/accessibility/radio-button-checkbox-size-expected.txt

Radio1:Pass
Checkbox1:Pass
Radio2:Pass
Checkbox2:Pass
Test
Test
This is a long test label that should be included in the size of this object
This is a long test label that should be included in the size of this object
